Fix VC++ warnings.

+++ include/ookii/usage_writer.h
@@ -845,7 +845,7 @@ namespace ookii
 
         //! \brief Writes the specified amount of spaces.
         //! \param count The number of spaces.
-        virtual void write_spacing(int count)
+        virtual void write_spacing(size_t count)
         {
             for (int i = 0; i < count; ++i)
             {
